backend/requirements.txt
flask==2.3.2
opencv-python-headless==4.7.0.72
deepface==0.0.72
requests==2.28.1
gunicorn==20.1.0
numpy==1.23.5
tensorflow==2.9.3
flask-cors==3.0.10
python-dotenv==1.0.0